feat: implemeting GenerateLabseqUseCase and InMemoryCacheControl

+++ src/main/java/com/interview/labseq/adpters/in/LabseqController.java
@@ -16,7 +16,8 @@ public class LabseqController {
     private GenerateLabseqUseCase useCase;
 
     @GetMapping
-    public int getLabseq(@PathVariable("n") int n){
+    public long getLabseq(@PathVariable("n") int n){
+
         return useCase.run(n);
     }
 }
